backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
pymysql==1.1.0
aiomysql==0.2.0
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# Redis/KVrocks client
redis==5.0.1

# Validation and serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
httpx==0.26.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
